Moved OpsuOptions inside Options, removed unused imports, other things in reaction to https://github.com/itdelatrisu/opsu/pull/8#issuecomment-55972945
