+++ src/itdelatrisu/opsu/GameScore.java
@@ -18,6 +18,8 @@
 
 package itdelatrisu.opsu;
 
+import itdelatrisu.opsu.states.Options.OpsuOptions;
+
 import java.io.File;
 import java.util.HashMap;
 import java.util.Iterator;
