+++ src/itdelatrisu/opsu/OsuParser.java
@@ -18,6 +18,8 @@
 
 package itdelatrisu.opsu;
 
+import itdelatrisu.opsu.states.Options.OpsuOptions;
+
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
