+++ src/itdelatrisu/opsu/SoundController.java
@@ -18,7 +18,7 @@
 
 package itdelatrisu.opsu;
 
-import itdelatrisu.opsu.states.Options;
+import itdelatrisu.opsu.states.Options.OpsuOptions;
 
 import java.io.IOException;
 import java.net.URL;
